Add Module stuff

+++ src/Core_System.cpp
@@ -27,7 +27,7 @@ time_t getCurrentTime()
     auto s_Now = std::chrono::system_clock::now();
     auto s_Usec = std::chrono::duration_cast<std::chrono::microseconds>(s_Now.time_since_epoch()).count();
 
-    g_tv.tv_sec = s_Usec / c_timeConst;
+    g_tv.tv_sec = static_cast<long>(s_Usec / c_timeConst);
     g_tv.tv_usec = s_Usec % c_timeConst;
 
     g_old.tv_sec = g_tv.tv_sec * c_timeConst + g_tv.tv_usec;
